1. Account에 대한 Get, Patch, Delete 기능 추가 및 테스트 작성

+++ src/main/kotlin/joowon/study/kotlin/accounts/AccountRepository.kt
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository
  * Time: 18:03
  */
 @Repository
-interface AccountRepository : JpaRepository<Account, Long> {
+interface AccountRepository : JpaRepository<Account, Int> {
     fun findByEmail(email: String) : Account?
 }
